requirements.txt
pytz~=2024.2
fastapi~=0.114
pydantic~=2.9
uvicorn~=0.30
python-dotenv~=1.0
tortoise-orm~=0.21